Попытался отвязать vm от контекста, который попадает в ее разные методы из фрагмента.
