+++ app/src/main/java/com/skysoft/skyweather/model/const.kt
@@ -6,6 +6,7 @@ const val IS_RUSSIAN_KEY = "IS_RUSSIAN_KEY"
 const val CITY_KEY = "CITY_KEY"
 const val ACTION_ON_LOAD_WEATHER = "ACTION_ON_LOAD_WEATHER"
 const val ACTION_ON_ERROR_LOAD_WEATHER = "ACTION_ON_ERROR_LOAD_WEATHER"
+const val ACTION_ON_ERROR_NO_INTERNET = "ACTION_ON_ERROR_NO_INTERNET"
 const val YANDEX_API_KEY = "X-Yandex-API-Key"
 const val YANDEX_API_URL = "https://api.weather.yandex.ru/"
 const val YANDEX_API_URL_END_POINT = "v2/informers"
