+++ app/src/main/java/com/skysoft/skyweather/utils/InternetService.kt
@@ -1,7 +1,12 @@
 package com.skysoft.skyweather.utils
 
 import android.app.IntentService
+import android.content.Context
 import android.content.Intent
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
+import com.skysoft.skyweather.R
 import com.skysoft.skyweather.model.*
 import com.skysoft.skyweather.repository.RepositoryImpl
 import retrofit2.Call
@@ -29,8 +34,33 @@ class InternetService(name: String = "") : IntentService(name) {
     }
 
     override fun onHandleIntent(intent: Intent?) {
-        intent?.getParcelableExtra<City>(CITY_KEY)?.let {
-            RepositoryImpl().getWeatherFromServer(it, callback)
+        if (hasInternet()) {
+            intent?.getParcelableExtra<City>(CITY_KEY)?.let {
+                RepositoryImpl().getWeatherFromServer(it, callback)
+            }
+        }else{
+            applicationContext.sendBroadcast(Intent(ACTION_ON_ERROR_NO_INTERNET).apply {
+                this.putExtra(ERROR_KEY, resources.getString(R.string.text_no_internet))
+            })
+        }
+    }
+
+    private fun hasInternet(): Boolean {
+        val connectivityManager =
+            applicationContext.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            val network = connectivityManager.activeNetwork ?: return false
+            val activeNetwork = connectivityManager.getNetworkCapabilities(network) ?: return false
+
+            return when {
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
+                else -> false
+            }
+        } else {
+            val networkInfo =
+                connectivityManager.activeNetworkInfo ?: return false
+            return networkInfo.isConnected
         }
     }
