+++ app/src/main/java/com/skysoft/skyweather/view/weathercard/WeatherViewModel.kt
@@ -9,7 +9,6 @@ import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.skysoft.skyweather.R
 import com.skysoft.skyweather.model.*
 import com.skysoft.skyweather.utils.InternetService
 import com.skysoft.skyweather.view.AppStateWeather
@@ -19,7 +18,6 @@ class WeatherViewModel(
     private val liveData: MutableLiveData<AppStateWeather> = MutableLiveData()
 ) : ViewModel() {
 
-    private var hasInternet = false
     private lateinit var cityToLoad: City
 
     fun getLiveData(): LiveData<AppStateWeather> {
@@ -28,16 +26,10 @@ class WeatherViewModel(
 
     private fun getWeatherFromServer(city: City, context: Context?) {
         context?.let {
-            hasInternet = checkForInternet(context)
             cityToLoad = city
-            if (!hasInternet) {
-                liveData.value =
-                    AppStateWeather.Error(context.resources.getString(R.string.text_no_internet))
-            } else {
-                context.startService(Intent(context, InternetService::class.java).apply {
-                    putExtra(CITY_KEY, city)
-                })
-            }
+            context.startService(Intent(context, InternetService::class.java).apply {
+                putExtra(CITY_KEY, city)
+            })
         }
     }
 
@@ -61,31 +53,15 @@ class WeatherViewModel(
                     liveData.value = AppStateWeather.SuccessLoadWeather(weatherDTO)
                 }
 
-            } else if (it.action == ACTION_ON_ERROR_LOAD_WEATHER) {
+            } else if (it.action == ACTION_ON_ERROR_LOAD_WEATHER ||
+                it.action == ACTION_ON_ERROR_NO_INTERNET
+            ) {
                 it.getStringExtra(ERROR_KEY)?.let { errorString ->
                     liveData.value = AppStateWeather.Error(errorString)
                 }
+
             } else {
             }
-        }
-    }
-
-    private fun checkForInternet(context: Context): Boolean {
-        val connectivityManager =
-            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            val network = connectivityManager.activeNetwork ?: return false
-            val activeNetwork = connectivityManager.getNetworkCapabilities(network) ?: return false
-
-            return when {
-                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
-                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
-                else -> false
-            }
-        } else {
-            val networkInfo =
-                connectivityManager.activeNetworkInfo ?: return false
-            return networkInfo.isConnected
         }
     }
